LayoutTests/platform/ios-simulator/fast/events/touch/gesture/gesture-scrollbar-expected.txt

CONSOLE MESSAGE: line 50: TypeError: undefined is not a function (evaluating 'eventSender.gestureTapDown(scrollbarX, scrollbarY)')
This tests scroll gesture events on a scroll bar. The scrollable div below should be slightly scrolled down if successful.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS movingDiv.scrollTop is 0
